requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.0.12
Django==3.2.15
django-bootstrap4==21.1
djangorestframework==3.14.0
idna==3.4
importlib-metadata==2.1.3
pytz==2022.2.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.12
zipp==3.6.0
